requirements.txt
dnspython==1.15.0
docker-py==1.10.6